+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/listeners/hover_listener.rb
@@ -8,11 +8,13 @@ module RubyLsp
       def initialize(response_builder, node_context, index, dispatcher)
         @response_builder = response_builder
         @handler = RubyLsp::Ree::HoverHandler.new(index, node_context)
+        @ree_context = RubyLsp::Ree::ReeContext.new(node_context)
 
         dispatcher.register(
           self, 
           :on_call_node_enter,
           :on_symbol_node_enter,
+          :on_string_node_enter, 
           :on_constant_read_node_enter
         )
       end
@@ -36,8 +38,23 @@ module RubyLsp
         $stderr.puts("error in hover listener(on_call_node_enter): #{e.message} : #{e.backtrace.first}")
       end
 
+      def on_string_node_enter(node)
+        return unless @ree_context.is_error_definition?
+        
+        hover_items = @handler.get_error_locales_hover_items(node)
+        
+        put_items_into_response(hover_items)
+      rescue => e
+        $stderr.puts("error in hover listener(on_string_node_enter): #{e.message} : #{e.backtrace.first}")
+      end
+
       def on_symbol_node_enter(node)
-        hover_items = @handler.get_linked_object_hover_items(node)
+        hover_items = if @ree_context.is_error_definition?
+          @handler.get_error_code_hover_items(node)
+        else
+          @handler.get_linked_object_hover_items(node)
+        end
+
         put_items_into_response(hover_items)
       rescue => e
         $stderr.puts("error in hover listener(on_symbol_node_enter): #{e.message} : #{e.backtrace.first}")