+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/utils/ree_lsp_utils.rb
@@ -7,7 +7,7 @@ module RubyLsp
         file_name = file_path + ".rb"
         Dir[File.join('**', file_name)].first
       end
-
+      
       def package_name_from_uri(uri)
         uri_parts = uri.to_s.split('/')
         
@@ -169,6 +169,17 @@ module RubyLsp
           names_with_commas = names.join(", ")
           :"(#{names_with_commas})"
         end
+      end
+
+      # copied from ree string_utils
+      def underscore(camel_cased_word)
+        return camel_cased_word unless /[A-Z-]|::/.match?(camel_cased_word)
+        word = camel_cased_word.to_s.gsub("::".freeze, "/".freeze)
+        word.gsub!(/([A-Z\d]+)([A-Z][a-z])/, '\1_\2'.freeze)
+        word.gsub!(/([a-z\d])([A-Z])/, '\1_\2'.freeze)
+        word.tr!("-".freeze, "_".freeze)
+        word.downcase!
+        word
       end
     end
   end